chore: split server.go
